clear things up
