+++ lib/august_db/memtable/memtable.ex
@@ -1,7 +1,7 @@
 defmodule Memtable do
   use Agent
 
-  defstruct current: :gb_trees.empty(), flushing: :gb_trees.empty()
+  defstruct [:current, :flushing]
 
   def start_link(initial_value) do
     Agent.start_link(fn -> initial_value end, name: __MODULE__)
@@ -9,28 +9,21 @@ defmodule Memtable do
 
   def query(key) do
     case Agent.get(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
-           case :gb_trees.lookup(key, current) do
-             :none -> :gb_trees.lookup(key, flushing)
+           case Memtable.Dirty.query(current, key) do
+             {:none, _} -> Memtable.Dirty.query(flushing, key)
              some -> some
            end
          end) do
-      {:value, {:value, data, time}} ->
-        {:value, data, time}
-
-      {:value, {:tombstone, time}} ->
-        {:tombstone, time}
-
-      :none ->
-        :none
+      # hide the empty string that we used in rust
+      {:tombstone, ""} -> :tombstone
+      {:none, ""} -> :none
+      value -> value
     end
   end
 
   def update(key, value) when is_binary(key) and is_binary(value) do
-    Agent.update(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
-      %__MODULE__{
-        current: :gb_trees.enter(key, {:value, value, System.monotonic_time()}, current),
-        flushing: flushing
-      }
+    Agent.get(__MODULE__, fn %__MODULE__{current: current, flushing: _flushing} ->
+      :ok = Memtable.Dirty.update(current, key, value)
     end)
 
     Memtable.Sizer.resize(key, value)
@@ -38,12 +31,9 @@ defmodule Memtable do
     :ok
   end
 
-  def delete(key) do
-    Agent.update(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
-      %__MODULE__{
-        current: :gb_trees.enter(key, {:tombstone, System.monotonic_time()}, current),
-        flushing: flushing
-      }
+  def delete(key) when is_binary(key) do
+    Agent.get(__MODULE__, fn %__MODULE__{current: current, flushing: _flushing} ->
+      :ok = Memtable.Dirty.delete(current, key)
     end)
 
     Memtable.Sizer.remove(key)
@@ -108,5 +98,59 @@ defmodule Memtable do
         flushing: :gb_trees.empty()
       }
     end)
+  end
+end
+
+defmodule Memtable.Dead do
+  use Agent
+
+  defstruct current: :gb_trees.empty(), flushing: :gb_trees.empty()
+
+  def start_link(initial_value) do
+    Agent.start_link(fn -> initial_value end, name: __MODULE__)
+  end
+
+  def query(key) do
+    case Agent.get(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
+           case :gb_trees.lookup(key, current) do
+             :none -> :gb_trees.lookup(key, flushing)
+             some -> some
+           end
+         end) do
+      {:value, {:value, data, time}} ->
+        {:value, data, time}
+
+      {:value, {:tombstone, time}} ->
+        {:tombstone, time}
+
+      :none ->
+        :none
+    end
+  end
+
+  def update(key, value) when is_binary(key) and is_binary(value) do
+    Agent.update(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
+      %__MODULE__{
+        current: :gb_trees.enter(key, {:value, value, System.monotonic_time()}, current),
+        flushing: flushing
+      }
+    end)
+
+    Memtable.Sizer.resize(key, value)
+
+    :ok
+  end
+
+  def delete(key) do
+    Agent.update(__MODULE__, fn %__MODULE__{current: current, flushing: flushing} ->
+      %__MODULE__{
+        current: :gb_trees.enter(key, {:tombstone, System.monotonic_time()}, current),
+        flushing: flushing
+      }
+    end)
+
+    Memtable.Sizer.remove(key)
+
+    :ok
   end
 end
